+++ src/js/renderers/additional/MaterialLabelRenderer.jsx
@@ -1,0 +1,47 @@
+/*
+  The MIT License
+
+  Copyright (c) 2017-2019 EclipseSource Munich
+  https://github.com/eclipsesource/jsonforms
+
+  Permission is hereby granted, free of charge, to any person obtaining a copy
+  of this software and associated documentation files (the "Software"), to deal
+  in the Software without restriction, including without limitation the rights
+  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  copies of the Software, and to permit persons to whom the Software is
+  furnished to do so, subject to the following conditions:
+
+  The above copyright notice and this permission notice shall be included in
+  all copies or substantial portions of the Software.
+
+  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+  THE SOFTWARE.
+*/
+import React from "react"
+import { rankWith, uiTypeIs } from "@jsonforms/core"
+import { withJsonFormsLabelProps } from "@jsonforms/react"
+import { Hidden, Typography } from "@mui/material"
+
+/**
+ * Default tester for a label.
+ * @type {RankedTester}
+ */
+export const materialLabelRendererTester = rankWith(3, uiTypeIs("Label"))
+
+/**
+ * Default renderer for a label.
+ */
+export const MaterialLabelRenderer = ({ text, visible }) => {
+  return (
+    <Hidden xsUp={!visible}>
+      <Typography variant="h6">{text}</Typography>
+    </Hidden>
+  )
+}
+
+export default withJsonFormsLabelProps(MaterialLabelRenderer)
