+++ webpack.config.js
@@ -45,12 +45,7 @@ module.exports = {
             options: { name: '[name].[ext]' } 
           }
         }, //for images
-        { test: /\.woff($|\?)|\.woff2($|\?)|\.ttf($|\?)|\.eot($|\?)|\.svg($|\?)/, use: ['file-loader'] }, //for fonts
-        {
-          test: /\.(ts|tsx)$/, // Add a rule for TypeScript files
-          exclude: /node_modules/,
-          use: 'ts-loader',
-        },
+        { test: /\.woff($|\?)|\.woff2($|\?)|\.ttf($|\?)|\.eot($|\?)|\.svg($|\?)/, use: ['file-loader'] } //for fonts
     ]
   },
   resolve: {
